Revert "fix: restructure commands and improve wording"
This reverts commit af52a4acac1684637f18283ce02fc5ca1cd1db9a.
