--- src/commands/bridge/command.ts
@@ -1,3 +1,0 @@
-import Program from "../../program.js";
-
-export default Program.command("bridge").description("Bridge operations for zkSync");
