--- src/commands/bridge/index.ts
@@ -1,5 +1,0 @@
-import "./deposit.js";
-import "./withdraw.js";
-import "./withdraw-finalize.js";
-
-import "./command.js"; // registers all the commands above
