--- src/commands/create/command.ts
@@ -1,3 +1,0 @@
-import Program from "../../program.js";
-
-export default Program.command("create").description("Scaffold new project for zkSync");
