--- src/commands/create/index.ts
@@ -1,3 +1,0 @@
-import "./create.js";
-
-import "./command.js"; // registers all the commands above
