+++ src/commands/dev/command.ts
@@ -1,3 +1,3 @@
 import Program from "../../program.js";
 
-export default Program.command("dev").description("Manage local zkSync development environment");
+export default Program.command("dev").description("All-in-one tool for local zkSync development");
