+++ src/commands/dev/start.ts
@@ -115,4 +115,4 @@ export const handler = async () => {
   }
 };
 
-Program.command("start").description("Start local zkSync environment and modules").action(handler);
+Program.command("start").description("Starts the local zkSync environment and modules").action(handler);
